refactor(core): update tests to use local validated helper
Add valibot as devDependency and create a test-scoped validated() helper
in src/test-helpers/ that mirrors the validate-valibot implementation.
This avoids a circular package dependency while keeping the test behavior
identical to the real validated() function.

Verified: precommit-verify ✓ (9a7985e90f8a01f2)
