+++ packages/core/package.json
@@ -53,6 +53,7 @@
   "devDependencies": {
     "@needle-di/core": "1.1.2",
     "@types/node": "22.19.17",
-    "neverthrow": "8.2.0"
+    "neverthrow": "8.2.0",
+    "valibot": "1.3.1"
   }
 }
